feat: Snowflake Compare
