+++ src/interactions/index.ts
@@ -2,4 +2,4 @@ export * from "./context/userinfoContext.js";
 export * from "./context/intentsLookupContext.js";
 export * from "./slash/userinfo.js";
 export * from "./slash/bitfieldLookup.js";
-export * from "./slash/snowflakeInfo.js";
+export * from "./slash/snowflakeCommands.js";
